backend/requirements.txt
asgiref==3.11.1
Django==5.2.12
django-cors-headers==4.9.0
djangorestframework==3.17.1
psycopg2-binary==2.9.11
sqlparse==0.5.5
typing_extensions==4.15.0
